refactor: remove shared ui type shim
