--- src/typescript/shared/ui/src/domains/GameEngineTypes.ts
@@ -1,1 +1,0 @@
-export * from "../components/GameEngineTypes";
