+++ src/typescript/shared/ui/src/domains/GameEngineUIService.ts
@@ -7,7 +7,7 @@
  * Follows Interface Segregation: exposes minimal, well-defined contract.
  */
 
-import type { ContextMenuAction, EngineStatus } from "./GameEngineTypes";
+import type { ContextMenuAction, EngineStatus } from "../components/GameEngineTypes";
 import { InputAggregator, type MovementInput } from "./InputDomain";
 import { InteractionManager, type InteractionOutcome } from "./InteractionDomain";
 import { type ViewportSize, ViewportSizer } from "./ViewportDomain";
